support log artifacts for ppl result logging
